pyproject.toml
[project]
name = "projeto-lab-prog"
version = "0.1.0"
description = ""
authors = [
    {name = "Franco Mansur Giani",email = "160537805+francomansur@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "playwright (>=1.52.0,<2.0.0)",
    "lxml (>=5.4.0,<6.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "openai (>=1.82.0,<2.0.0)",
    "replicate (>=1.0.6,<2.0.0)",
    "huggingface-hub (>=0.32.0,<0.33.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
